KeyEvents and code improvements
Got rid off the slave_respond enumeration. Now the slave responds with the code of command if it was succesful, or fail code if it was not.
Also added keyEvents to control r g b leds. For example R key turns on the red led, and turns off when key released.